backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.60
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.13
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.2
duckduckgo-search==3.9.6